Remove nav items on 'bp_actions' at priority 9.
We need to run our nav removal routine before priority 10 to fix a
race condition with bp_is_group() becoming true on the group creation
page and causing a fatal error.

The fatal error occurs because we are attempting to load the group
nav, which is only accessible on a single group page.